// src/enums/Sockets.ts
export enum SocketEvents {
	ServerRequestQueue = 'server.requestQueue',
	ServerRequestPlayTrack = 'server.requestPlayTrack',
	ServerRequestPreviousTrack = 'server.requestPreviousTrack',
	ServerRequestSkipTrack = 'server.requestSkipTrack',
	ServerRequestRemoveTrack = 'server.requestRemoveTrack',
	ServerRequestPlayPlaylist = 'server.requestPlayPlaylist',
	ServerRequestResumeMusicPlayer = 'server.requestResumeMusicPlayer',
	ServerRequestPauseMusicPlayer = 'server.requestPauseMusicPlayer',
	ServerRequestedQueue = 'server.requestedQueue',
	ServerPlayedTrack = 'server.playedTrack',
	ServerSkipedTrack = 'server.skipedTrack',
	ServerRemovedTrack = 'server.removedTrack',
	ServerResumedMusicPlayer = 'server.resumedMusicPlayer',
	ServerPausedMusicPlayer = 'server.pausedMusicPlayer',
	ServerDeletedQueue = 'server.deletedQueue',
	ServerStartedPlaying = 'server.startedPlaying',
	ServerEmptyQueue = 'server.emptyQueue',
	ServerConnectionDestroyed = 'server.connectionDestroyed',
	BotRequestedQueue = 'bot.requestedQueue',
	BotPlayedTrack = 'bot.playedTrack',
	BotSkipedTrack = 'bot.skipedTrack',
	BotRemovedTrack = 'bot.removedTrack',
	BotResumedMusicPlayer = 'bot.resumedMusicPlayer',
	BotPausedMusicPlayer = 'bot.pausedMusicPlayer',
	BotDeletedQueue = 'bot.deletedQueue',
	BotStartedPlaying = 'bot.startedPlaying',
	BotEmptyQueue = 'bot.emptyQueue',
	BotConnectionDestroyed = 'bot.connectionDestroyed',
	BotError = 'bot.error'
}